+++ packages/shared/src/dir/index.ts
@@ -8,3 +8,4 @@ export {
 } from './system';
 export {streamingListEntries} from './list';
 export {WorkspaceFileStructure, TreeifyResult} from './tree';
+export {projectConfig, projectCompiledConfig, projectRulesDirectory, projectRules} from './project';